+++ src/engine/core/CPlanes.h
@@ -17,9 +17,9 @@ public:
 	unsigned char getbit(unsigned char plane);
 	void setOffsets(unsigned long p1, unsigned long p2, unsigned long p3,
 					 unsigned long p4, unsigned long p5 = 0);
-	void readPlane(size_t p, Uint8 *pixels, Uint16 width, Uint16 height);
-	void readPlaneofTiles(size_t p, Uint8 *pixels, Uint16 columns,
-								Uint16 tilesize, Uint16 numtiles);
+    void readPlane(uint32_t p, uint8_t *pixels, uint16_t width, uint16_t height);
+    void readPlaneofTiles(uint32_t p, uint8_t *pixels, uint16_t columns,
+                                uint16_t tilesize, uint16_t numtiles);
 	
 private:
 	unsigned long getbit_bytepos[5];
